Croquizador/requirements.txt
flask
requests
lxml
Pillow

